# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Logistics Optimization Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /yard/schedule-arrival:
    post:
      operationId: scheduleArrival
      tags: [yard]
      summary: Schedule a truck arrival
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ScheduleArrivalBody"
      responses:
        "201":
          description: Scheduled arrival
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TruckArrival"
  /yard/assign-dock:
    post:
      operationId: assignDock
      tags: [yard]
      summary: Assign a truck to a dock
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignDockBody"
      responses:
        "200":
          description: Dock assignment result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssignDockResult"
...
  /workers:
    get:
      operationId: getWorkers
      tags: [workforce]
      summary: List all workers
      responses:
        "200":
          description: Workers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Worker"
    post:
      operationId: createWorker
      tags: [workforce]
      summary: Add a worker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateWorkerBody"
      responses:
        "201":
          description: Created worker
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Worker"
  /workers/assign-task:
    post:
      operationId: assignWorkerTask
      tags: [workforce]
      summary: Assign a task to a worker
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignTaskBody"
      responses:
        "200":
          description: Task created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkerTask"
...
components:
  schemas:
...
    TruckArrival:
      type: object
      properties:
        id:
          type: string
        truck_id:
          type: string
        scheduled_time:
          type: string
        actual_arrival:
          type: string
          nullable: true
        assigned_slot:
          type: string
          nullable: true
        dock_id:
          type: string
          nullable: true
        status:
          type: string
          enum: [scheduled, arrived, docked, departed]
      required: [id, truck_id, scheduled_time, status]
...
    ScheduleArrivalBody:
      type: object
      properties:
        truck_id:
          type: string
        scheduled_time:
          type: string
      required: [truck_id, scheduled_time]
    AssignDockBody:
      type: object
      properties:
        truck_id:
          type: string
        dock_id:
          type: string
      required: [truck_id, dock_id]
    AssignDockResult:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        dock_id:
          type: string
        estimated_wait_minutes:
          type: number
      required: [success, message, dock_id, estimated_wait_minutes]

    # ── Fragmentation ─────────────────────────────────────────────────────
...
    Worker:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [loader, driver, supervisor, scanner, forklift_operator]
        warehouse_id:
          type: string
        performance_score:
          type: number
        reward_points:
          type: integer
        tasks_completed:
          type: integer
        status:
          type: string
          enum: [active, on_break, off_duty]
        ai_recommendation:
          type: string
          nullable: true
      required: [id, name, role, warehouse_id, performance_score, reward_points, tasks_completed, status]
    WorkerTask:
      type: object
      properties:
        id:
          type: string
        worker_id:
          type: string
        task_type:
          type: string
        description:
          type: string
        assigned_time:
          type: string
        completed:
          type: boolean
        completed_time:
          type: string
          nullable: true
        priority:
          type: string
          enum: [low, medium, high]
      required: [id, worker_id, task_type, description, assigned_time, completed, priority]
    CreateWorkerBody:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
          enum: [loader, driver, supervisor, scanner, forklift_operator]
        warehouse_id:
          type: string
      required: [name, role, warehouse_id]
    AssignTaskBody:
      type: object
      properties:
        worker_id:
          type: string
        task_type:
          type: string
        description:
          type: string
        priority:
          type: string
          enum: [low, medium, high]
      required: [worker_id, task_type, description]
...